requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-import-export==4.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
